Port upstream libpng 1.6.40
This work is based off libpng 1.6.40 autoconfigured on a Linux x86_64
system. Header files under `include/` are autogenerated and adapted to
unikraft as needed. Hardware acceleration is available on x86_64
automatically if the intrinsics library is enabled and the required CPU
features are present. The `pngtest` program is provided as an optional
main function.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #3